+++ integration-tests/gatsby-cli/test-helpers/invoke-cli.js
@@ -2,6 +2,7 @@ import execa, { sync } from "execa"
 import { join } from "path"
 import strip from "strip-ansi"
 import { createLogsMatcher } from "./matcher"
+import kill from "tree-kill"
 
 const gatsbyBinLocation = join(
   __dirname,
@@ -11,18 +12,38 @@ const gatsbyBinLocation = join(
   `cli.js`
 )
 
+function waitChildProcessExit(pid, resolve, reject, attempt = 0) {
+  try {
+    process.kill(pid, 0) // check if process is still running
+    if (attempt > 15) {
+      reject(new Error("Gatsby process hasn't exited in 15 seconds"))
+      return
+    }
+    setTimeout(() => {
+      waitChildProcessExit(pid, resolve, reject, attempt + 1)
+    }, 1000)
+  } catch (e) {
+    resolve()
+  }
+}
 // Use as `GatsbyCLI.cwd('execution-folder').invoke('new', 'foo')`
 export const GatsbyCLI = {
   from(relativeCwd) {
     return {
       invoke(args) {
-        const NODE_ENV = args[0] === `develop` ? `development` : `production`
+        const NODE_ENV =
+          (Array.isArray(args) ? args[0] : args) === `develop`
+            ? `development`
+            : `production`
         try {
-          const results = sync("node", [gatsbyBinLocation].concat(args), {
-            cwd: join(__dirname, `../`, `./${relativeCwd}`),
-            env: { NODE_ENV, CI: 1, GATSBY_LOGGER: `ink` },
-          })
-
+          const results = sync(
+            process.execPath,
+            [gatsbyBinLocation].concat(args),
+            {
+              cwd: join(__dirname, `../`, `./${relativeCwd}`),
+              env: { NODE_ENV, CI: 1, GATSBY_LOGGER: `ink` },
+            }
+          )
           return [
             results.exitCode,
             createLogsMatcher(strip(results.stdout.toString())),
@@ -36,10 +57,27 @@ export const GatsbyCLI = {
       },
 
       invokeAsync: (args, onExit) => {
-        const NODE_ENV = args[0] === `develop` ? `development` : `production`
-        const res = execa("node", [gatsbyBinLocation].concat(args), {
+        const NODE_ENV =
+          (Array.isArray(args) ? args[0] : args) === `develop`
+            ? `development`
+            : `production`
+        const res = execa(process.execPath, [gatsbyBinLocation].concat(args), {
           cwd: join(__dirname, `../`, `./${relativeCwd}`),
           env: { NODE_ENV, CI: 1, GATSBY_LOGGER: `ink` },
+        })
+
+        let isKilled = false
+        const onExitPromise = new Promise((resolve, reject) => {
+          res.on(`exit`, () => {
+            // give it some time to exit
+            waitChildProcessExit(res.pid, resolve, reject)
+          })
+
+          res.catch(err => {
+            if (!isKilled) {
+              reject(err)
+            }
+          })
         })
 
         let logs = ``
@@ -48,21 +86,13 @@ export const GatsbyCLI = {
             logs += data.toString()
           }
 
-          if (onExit && onExit(strip(logs))) {
-            res.cancel()
+          if (!isKilled && onExit && onExit(strip(logs))) {
+            isKilled = true
+            kill(res.pid, "SIGINT")
           }
         })
 
-        return [
-          res.catch(err => {
-            if (err.isCanceled) {
-              return
-            }
-
-            throw err
-          }),
-          () => createLogsMatcher(strip(logs)),
-        ]
+        return [onExitPromise, () => createLogsMatcher(strip(logs))]
       },
     }
   },