+++ packages/gatsby-plugin-utils/src/polyfill-remote-file/__tests__/gatsby-image-resolver.ts
@@ -681,6 +681,33 @@ describe(`gatsbyImageData`, () => {
     `)
   })
 
+  it(`should generate tracedSVG placeholder`, async () => {
+    fetchRemoteFile.mockResolvedValueOnce(
+      path.join(__dirname, `__fixtures__`, `dog-portrait.jpg`)
+    )
+    const fixedResult = await gatsbyImageResolver(
+      portraitSource,
+      {
+        layout: `fixed`,
+        width: 300,
+        placeholder: PlaceholderType.TRACED_SVG,
+      },
+      actions
+    )
+
+    expect(fetchRemoteFile).toHaveBeenCalledTimes(1)
+    expect(fetchRemoteFile).toHaveBeenCalledWith({
+      url: portraitSource.url,
+      cacheKey: `1`,
+      directory: expect.stringContaining(cacheDir),
+    })
+    expect(fixedResult?.placeholder).toMatchInlineSnapshot(`
+      Object {
+        "fallback": "data:image/svg+xml,%3csvg%20xmlns='http://www.w3.org/2000/svg'%20width='20'%20height='33'%20viewBox='0%200%2020%2033'%3e%3cpath%20d='M6%201C4%205%204%205%203%203L2%201C0%201%200%205%200%2014s0%209%203%208c4%200%204%200%204-2v-8H6c0-1%202-3%204-3s2%200%202-2l-1-4c0-3%200-3-3-3L6%201'%20fill='%23d3d3d3'%20fill-rule='evenodd'/%3e%3c/svg%3e",
+      }
+    `)
+  })
+
   it(`should render avif, webp other format in this order`, async () => {
     const constrainedResult = await gatsbyImageResolver(
       portraitSource,
@@ -695,6 +722,8 @@ describe(`gatsbyImageData`, () => {
 
     expect(constrainedResult?.images.sources[0].type).toBe(`image/avif`)
     expect(constrainedResult?.images.sources[1].type).toBe(`image/webp`)
+    expect(constrainedResult?.images.fallback.src).toContain(`dog-portrait.jpg`)
+
     expect(constrainedResult?.images.sources.length).toBe(2)
   })
 })